requirements.txt
google.generativeai 
streamlit
pypdf
Documents
opencv-python
python-docx
pillow 
 
